organize local preferences and migrate announcements feature
