+++ app/src/main/java/edu/rpi/shuttletracker/app/di/ShuttleTrackerModule.kt
@@ -16,10 +16,10 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import edu.rpi.shuttletracker.BuildConfig
+import edu.rpi.shuttletracker.data.local.preferences.UserPreferencesRepository
 import edu.rpi.shuttletracker.data.remote.ShuttleApi
 import edu.rpi.shuttletracker.data.remote.dto.RouteDto
 import edu.rpi.shuttletracker.data.remote.dto.RouteDtoDeserializer
-import edu.rpi.shuttletracker.data.repository.UserPreferencesRepository
 import edu.rpi.shuttletracker.util.FlattenTypeAdapterFactory
 import edu.rpi.shuttletracker.util.hasNetwork
 import kotlinx.coroutines.CoroutineScope
