+++ app/src/main/java/edu/rpi/shuttletracker/feature/settings/analytics/AnalyticsViewModel.kt
@@ -3,8 +3,8 @@ package edu.rpi.shuttletracker.feature.settings.analytics
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import edu.rpi.shuttletracker.data.local.preferences.UserPreferencesRepository
 import edu.rpi.shuttletracker.data.models.AnalyticsFactory
-import edu.rpi.shuttletracker.data.repository.UserPreferencesRepository
 import kotlinx.coroutines.launch
 import javax.inject.Inject
