+++ app/src/main/java/edu/rpi/shuttletracker/feature/setup/SetupScreenViewModel.kt
@@ -3,7 +3,7 @@ package edu.rpi.shuttletracker.feature.setup
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import edu.rpi.shuttletracker.data.repository.UserPreferencesRepository
+import edu.rpi.shuttletracker.data.local.preferences.UserPreferencesRepository
 import kotlinx.coroutines.launch
 import javax.inject.Inject
